changed the API definition
Added the CreateSA field to the spec of our api.

+++ internal/controller/rbacrule_controller.go
@@ -170,7 +170,7 @@ func (r *RBACRuleReconciler) Reconcile(ctx context.Context, req ctrl.Request) (c
 		period := time.Until(end)
 		r.Log.Info("Rule will be scheduled for deletion", "Time until deletion", period)
 		return ctrl.Result{RequeueAfter: period}, nil
-	} else if end.Before(time.Now()) {
+	} else if end != (time.Time{}) && end.Before(time.Now()) {
 		err := r.Delete(ctx, RBACRule)
 		if err != nil {
 			r.Log.Error(err, "error deleting resource")
